feat: challenge suspicious requests before proxying them (#15)
`ProxyServe` now uses the `Tollkeeper` to check incoming requests before proxying them. The visa is being passed as a "X-Keeper-Visa" token which contains a base64-encoded json with the `Visa` info. (I am aware that the structures are still forgeable, this will be adressed in another PR).

Also read a bit more into multi threading with rust and overhauled the design to avoid broad locks (and complicated field definitions :smile)

~Still WIP:~

- [x]  Manual test of access with Visa

Reviewed-on: https://gitea.ascendise.ch/ascendise/tollkeeper/pulls/15
Co-authored-by: Ascendise <rapperido@gmail.com>
Co-committed-by: Ascendise <rapperido@gmail.com>