feat: sign visas and tolls (#17)
added a new wrapper type `Signed` that defines a structure with an HMAC-SHA256 signature of the wrapped value.

`Tollkeeper` now only accepts and returns `Signed<Toll>` and  `Signed<Visa>`.

This allows us to trust the structure returned by the user and keeps us from having to store state of visas server side.

Reviewed-on: https://gitea.ascendise.ch/ascendise/tollkeeper/pulls/17
Co-authored-by: Ascendise <rapperido@gmail.com>
Co-committed-by: Ascendise <rapperido@gmail.com>